:lipstick: Style: 그리드 사이 gap
gap 조절
